+++ src/PushNotifications.Api/Startup.cs
@@ -10,11 +10,11 @@ namespace PushNotifications.Api
     {
         public void Configuration(IAppBuilder app)
         {
-            log4net.Config.XmlConfigurator.Configure();
-
             var appContext = new ApplicationContext("PushNotifications");
             var cfgRepo = new ConsulForPandora(new Uri("http://consul.local.com:8500"));
             var pandora = new Pandora(appContext, cfgRepo);
+
+            LogStartup.Boot(pandora);
 
             app.UseHttpWebApi(pandora);
         }