+++ src/PushNotifications.Contracts/Subscriptions/Commands/UnSubscribeFromPushy.cs
@@ -1,0 +1,35 @@
+﻿using Elders.Cronus.DomainModeling;
+using System.Runtime.Serialization;
+
+namespace PushNotifications.Contracts.Subscriptions.Commands
+{
+    [DataContract(Name = "091df3c5-d619-4e4b-a0d6-596e1a5acb3e")]
+    public class UnSubscribeFromPushy : ICommand
+    {
+        UnSubscribeFromPushy() { }
+
+        public UnSubscribeFromPushy(PushySubscriptionId id, string userId, string token)
+        {
+            Id = id;
+            UserId = userId;
+            Token = token;
+        }
+
+        [DataMember(Order = 1)]
+        public PushySubscriptionId Id { get; private set; }
+
+        [DataMember(Order = 2)]
+        public string UserId { get; private set; }
+
+        [DataMember(Order = 3)]
+        public string Token { get; private set; }
+
+        public bool IsValid()
+        {
+            return
+               Id != null &&
+               !string.IsNullOrWhiteSpace(UserId) &&
+               !string.IsNullOrWhiteSpace(Token);
+        }
+    }
+}