+++ src/PushNotifications.Api/RELEASE_NOTES.md
@@ -1,3 +1,6 @@
+#### 2.0.10 - 06.02.2017
+* Add integration with Pushy push notifications service.
+
 #### 2.0.9 - 21.21.2016
 * Fixes throttling config